Ignore errors in parsing single properties
Sometimes a single property will fail to parse but the rest of the
event properties are valuable - it is important to keep going so we
can get some data back if possible. An example for this is the
provider {22fb2cd6-0e7b-422b-a0c7-2fad1fd0e716}
Microsoft-Windows-Kernel-Process with event id 1 (process
create). Without this change it fails to return any event properties
but actually it just fails to parse the flags.